+++ doc/changelog.rst
@@ -8,6 +8,8 @@ Changed
 Added
 ^^^^^
 - Implement :meth:`~scim2_client.SCIMClient.delete` `check_response_payload` attribute.
+- :class:`~scim2_models.ServiceProviderConfig`, :class:`~scim2_models.ResourceType`
+  and :class:`~scim2_models.Schema` are added to the default resource types list.
 
 Fixed
 ^^^^^